Complete CustomerController

+++ src/main/java/vn/com/assistant/fqcbackend/service/CustomerServiceImp.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import vn.com.assistant.fqcbackend.dto.CustomerRequestDTO;
 import vn.com.assistant.fqcbackend.dto.CustomerResponseDTO;
 import vn.com.assistant.fqcbackend.entity.Customer;
+import vn.com.assistant.fqcbackend.exception.ConflictException;
 import vn.com.assistant.fqcbackend.exception.InvalidException;
 import vn.com.assistant.fqcbackend.repository.CustomerRepository;
 import vn.com.assistant.fqcbackend.utility.CustomerMapper;
@@ -43,7 +44,9 @@ public class CustomerServiceImp implements CustomerService {
 
     @Override
     public void delete(String customerId) {
-        customerRepository.findById(customerId).orElseThrow(() -> new InvalidException(env.getProperty("customer.notExisted")));
+        Customer customer = customerRepository.findById(customerId).orElseThrow(() -> new InvalidException(env.getProperty("customer.notExisted")));
+        if (!(customer.getFabricList().isEmpty() && customer.getColorList().isEmpty()))
+            throw new ConflictException(env.getProperty("customer.used"));
         customerRepository.deleteById(customerId);
     }
 }
